Move all custom schema deserialization code into a schema/deserialization
